refactoring modal scss files into one. save before deleting the individual scss files.
