+++ src/__tests__/Action.test.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import { ReactTestCaseContext } from '../context/reducers/reactTestCaseReducer';
@@ -35,12 +34,13 @@ const mockDataState = {
 describe('Assertion ', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div');
-    ReactDOM.render(
+    render(
       <ReactTestCaseContext.Provider value={[reactTestCaseState, dispatchToReactTextCase]}>
         <MockDataContext.Provider value={[mockDataState, dispatchToMockData]}>
           <Action {...reactTestCaseState} />
         </MockDataContext.Provider>
-      </ReactTestCaseContext.Provider>, div,
+      </ReactTestCaseContext.Provider>,
+      div
     );
   });
 });
