+++ src/components/TestHelpModals/ReactHelpModal.jsx
@@ -4,7 +4,7 @@ import { GlobalContext } from '../../context/reducers/globalReducer';
 import { openBrowserDocs } from '../../context/actions/globalActions';
 import { closeInfoModal } from '../../context/actions/reactTestCaseActions';
 import { ReactTestCaseContext } from '../../context/reducers/reactTestCaseReducer';
-import styles from '../../components/TestHelpModals/ReactHelpModal.module.scss';
+import styles from '../../components/TestHelpModals/TestHelpModal.module.scss';
 import cn from 'classnames';
 
 const closeIcon = require('../../assets/images/close.png');
@@ -43,14 +43,17 @@ const ReactHelpModal = () => {
         onClick={closeModal}
         className={cn('far fa-window-close', styles.describeClose)}
       />
-      <h2>Describe(name, fn)</h2>
+      <h2>describe(name, fn)</h2>
       <p>
-        Describe creates a block that groups together several related tests. The name argument is
-        simply the name of component you're testing. fn argument is the test callback function{' '}
+        Describe creates a block that groups together several related tests. First argument is name
+        of component your testing. Second argument is the test/it callback function{' '}
       </p>
-      <p>Testing a new paragraph</p>
+      <h2>it(name, fn)</h2>
+      <p>
+        First argument is name of the test. Second argument is callback function with test method.{' '}
+      </p>
       <img className={styles.describe} src={describe} />
-      <img src={describe} />
+
       <a onClick={openDocs}>Need More Help?</a>
     </ReactModal>
   );
