+++ src/components/TestHelpModals/TestHelpModal.module.scss
@@ -50,6 +50,28 @@ a {
   }
 }
 
+// .modal {
+//   position: fixed;
+//   top: 35%;
+//   left: 30%;
+//   width: 500px;
+//   height: 250px;
+//   font-family: $openSans;
+//   background-color: white;
+//   border: 4px solid $mint;
+//   border-radius: 5px;
+//   box-shadow: 0 0 0 1px white inset;
+// }
+// h2 {
+//   font-size: 1.5em;
+//   font-weight: bold;
+//   padding: 5px;
+// }
+
+// p {
+//   padding: 5px;
+// }
+
 // #title {
 //   display: flex;
 //   justify-content: space-between;
